Adds changes from review
