+++ src/network/Network.consts.ts
